lenght

+++ 0x05-pointers_arrays_strings/2-strlen.c
@@ -6,7 +6,7 @@
  *
  * Return: Length.
  */
-size_t _strlen(const char *str)
+int _strlen(const char *str)
 {
 	len = 0;
